python_producer/pyproject.toml
[tool.poetry]
name = "python_producer"
version = "0.1.0"
description = "Python Apache Kafka producer"
authors = ["."]

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
main="producer:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
